[#178] refactor: api/modules/jjikboul.ts를 domains로 통합
- jjikboulApi에 getJjikboulDetail 추가
- JjikboulDetailResponse 타입 정의
- transform 함수 수정 (Level 필드명 수정)
- useGetJjikboulDetailQuery를 domains/jjikboul/hooks로 이동
- jjikboul page import 경로 및 필드명 수정
- useJjikboul 타입 수정
- api/modules/jjikboul.ts 제거
- api/hooks/jjikboul.ts 제거
