+++ src/domains/jjikboul/types/response.ts
@@ -1,26 +1,11 @@
+import type { MemberInfoResponse } from '@/domains/auth/types/response';
+import type { GymResponse, LevelResponse } from '@/domains/record/types/response';
+
 export interface JjikboulResponse {
   jjikboulId: number;
   problemType: string;
   description: string;
   problemUrl?: string;
-}
-
-export interface MemberInfoResponse {
-  memberId: number;
-  nickname: string;
-  profileUrl: string;
-}
-
-export interface GymResponse {
-  gymId: number;
-  name: string;
-  address: string;
-}
-
-export interface LevelResponse {
-  levelId: number;
-  name: string;
-  color: string;
 }
 
 export interface JjikboulDetailResponse {
